+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to the "much-assembly-required--upload-on-save" extension wi
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
 
+## [0.0.5] 2018-01-03
+
+Fix keybinding issue misnamed command
+
+## [0.0.4] 2017-12-31
+
+Add upload on save configuration option
+
 ## [0.0.3] 2017-12-31
 
 Attempt at error handling, improve readme